# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /tasks:
    get:
      tags:
        - task-entity-controller
      description: get-task
      operationId: getCollectionResource-task-get
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  - $ref: '#/components/schemas/CollectionModelEntityModelTask'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelTask'
            text/uri-list:
              schema:
                type: string
    put:
      tags:
        - task-controller
      operationId: updateTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Task'
    post:
      tags:
        - task-entity-controller
      description: create-task
      operationId: postCollectionResource-task-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTask'
  /tasks/{id}:
    get:
      tags:
        - task-entity-controller
      description: get-task
      operationId: getItemResource-task-get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTask'
        '404':
          description: Not Found
    put:
      tags:
        - task-entity-controller
      description: update-task
      operationId: putItemResource-task-put
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTask'
        '201':
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTask'
        '204':
          description: No Content
    delete:
      tags:
        - task-entity-controller
      description: delete-task
      operationId: deleteItemResource-task-delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
    patch:
      tags:
        - task-entity-controller
      description: patch-task
      operationId: patchItemResource-task-patch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelTask'
        '204':
          description: No Content
...
  /auth/signup:
    post:
      tags:
        - auth-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /auth/signin:
    post:
      tags:
        - auth-controller
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                type: string
  /tasks/{taskId}:
    get:
      tags:
        - task-controller
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      tags:
        - task-controller
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                type: string
components:
  schemas:
...
    TaskState:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        state:
          $ref: '#/components/schemas/TaskState'
        priority:
          type: string
        description:
          type: string
        type:
          type: string
        assigned:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        enabled:
          type: boolean
        username:
          type: string
        authorities:
          type: array
          items:
            $ref: '#/components/schemas/GrantedAuthority'
        credentialsNonExpired:
          type: boolean
        accountNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
    EntityModelTask:
      type: object
      properties:
        priority:
          type: string
        description:
          type: string
        type:
          type: string
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelTask:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelTask'
        _links:
          $ref: '#/components/schemas/Links'
...
    TaskRequestBody:
      type: object
      properties:
        id:
          type: integer
          format: int64
        state:
          type: string
        priority:
          type: string
        description:
          type: string
        type:
          type: string
        assigned:
          type: string
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'